requirements.txt
boto3==1.7.50
botocore==1.10.50
cachetools==2.1.0
dateutils==0.6.6
docutils==0.14
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
httplib2==0.11.3
jmespath==0.9.3
kubernetes==6.0.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
uritemplate==3.0.0
